lib/lo/xml#from_js plus first request body transform
for PUT /appliances/:app/configuration.

also, tests/tools#parseJSON mirrors parseXML.

BTW, had to squash this in after rebasing on top of the slms
block addition:

GET /appliances/:app/configuration: slms may be missing

the failing test creates a JS object from a JSON file, converts
it into XML using the PUT endpoint (that's the "real" test), and
then tries to convert it back to JS with the GET endpoint for
good measure.  but the GET side expects full configuration in XML
while the PUT side permits partial requests (those configuration
subtrees which are present get updated, the ones missing from
the request are left as is).

i don't have time for a proper fix ATM, so here's a quick hack:
the GET /appliances/:app/configuration endpoint treats the slms
block as optional.